examen_api/services/logbook_service/requirements.txt
pymongo[srv]
pydantic[email]
uvicorn
fastapi
python-dotenv